requirements.txt
Flask==1.1.2
gevent==20.9.0
Flask-MySQLdb==0.2.0
mysqlclient==2.0.1
mysql-connector-python
mysql-connector